Revert "Refactoring of Format data (#66)"
This reverts commit d853b886f444dbb7b35be7a4e27b481fa3f10ef1.
