+++ .gitignore
@@ -8,4 +8,3 @@
 *.pyc
 *.csv
 !tests/test_files/*.json
-*.egg-info
