+++ archive/fetch_fcd_data.py
@@ -4,7 +4,6 @@ Will create a json file for each request.
 
 import requests
 import json
-import os
 
 
 def save_json(json_data: dict, id: int) -> None:
@@ -53,7 +52,6 @@ def run_requests(start_url: str) -> None:
 
 
 def main():
-    API_KEY = os.getenv("FCD_API_KEY")
     url = f"https://fcd-share.civil.aau.dk/api/linestrings/?year=2014&format=json&apikey={API_KEY}"
     run_requests(url)
