+++ speed_limit_floating_car_data.egg-info/PKG-INFO
@@ -1,0 +1,3 @@
+Metadata-Version: 2.1
+Name: speed-limit-floating-car-data
+Version: 1.0
