+++ speed_limit_floating_car_data.egg-info/SOURCES.txt
@@ -1,0 +1,21 @@
+README.md
+setup.py
+archive/__init__.py
+archive/analysis_metrics.py
+archive/convert_json_to_df.py
+archive/fetch_fcd_data.py
+pipeline/__init__.py
+pipeline/preprocessing/__init__.py
+pipeline/preprocessing/df_processing.py
+pipeline/preprocessing/format_data.py
+pipeline/preprocessing/process_merged_geojson.py
+pipeline/preprocessing/compute_features/__init__.py
+pipeline/preprocessing/compute_features/calculate_speeds_distances.py
+pipeline/preprocessing/compute_features/compute_features.py
+pipeline/preprocessing/compute_features/feature.py
+pipeline/preprocessing/compute_features/type_alias.py
+pipeline/preprocessing/compute_features/vcr_calculator.py
+speed_limit_floating_car_data.egg-info/PKG-INFO
+speed_limit_floating_car_data.egg-info/SOURCES.txt
+speed_limit_floating_car_data.egg-info/dependency_links.txt
+speed_limit_floating_car_data.egg-info/top_level.txt
