+++ tests/preprocessing/df_processing_test.py
@@ -1,8 +1,7 @@
 import pytest
-import json
 import os
 import pandas as pd  # type: ignore
-from pipeline.preprocessing.formatting.FCD_Formatter import _remove_fcd_request_wrapper
+from pipeline.preprocessing.df_processing import create_df_from_json
 
 
 def test_create_df_from_json():
@@ -22,7 +21,5 @@ def test_create_df_from_json():
     }
     expected_df = pd.DataFrame(data=expected_data)
     testfile_path = os.getcwd() + "/tests/test_files/geo_json_trip_data.json"
-    with open(testfile_path, "r") as testfile:
-        data = json.load(testfile)
-        actual_df = _remove_fcd_request_wrapper(data)
-        assert expected_df.equals(actual_df)
+    actual_df = create_df_from_json(testfile_path)
+    assert expected_df.equals(actual_df)
